@@ -1,6 +1,6 @@
 package org.kaerdan.mvp_navigation;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import java.util.Arrays;
 import java.util.List;
