@@ -2,7 +2,7 @@ package org.kaerdan.mvp_navigation.core.ui.article;
 
 import org.kaerdan.mvp_navigation.core.data.Article;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 public interface ArticleContract {
     interface Presenter extends org.kaerdan.mvp_navigation.core.Presenter<View> { }
