@@ -5,9 +5,9 @@ import org.kaerdan.mvp_navigation.core.data.Article;
 
 import android.os.Bundle;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 
 import android.view.LayoutInflater;
 import android.view.View;
