@@ -3,7 +3,7 @@ package org.kaerdan.mvp_navigation.core.ui.favorite_list;
 import org.kaerdan.mvp_navigation.core.data.DataProvider;
 import org.kaerdan.mvp_navigation.core.ui.OnArticleClickListener;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 public class FavoriteListPresenter implements FavoriteListContract.Presenter, OnArticleClickListener {
 
