@@ -8,7 +8,7 @@ import android.content.Intent;
 
 import android.os.Bundle;
 
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 
 import android.view.MenuItem;
 
