@@ -3,16 +3,11 @@ package org.kaerdan.mvp_navigation.example3_viewpager;
 import org.kaerdan.mvp_navigation.R;
 import org.kaerdan.mvp_navigation.core.ui.article_list.ArticleListContract;
 import org.kaerdan.mvp_navigation.core.ui.favorite_list.FavoriteListContract;
-
 import android.os.Bundle;
-
-import android.support.annotation.NonNull;
-
-import android.support.design.widget.TabLayout;
-
-import android.support.v4.view.ViewPager;
-
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.NonNull;
+import androidx.viewpager.widget.ViewPager;
+import androidx.appcompat.app.AppCompatActivity;
+import com.google.android.material.tabs.TabLayout;
 
 public class ViewPagerActivity extends AppCompatActivity implements ArticleListContract.NavigatorProvider,
     FavoriteListContract.NavigatorProvider {
