@@ -4,7 +4,7 @@ import org.kaerdan.mvp_navigation.R;
 
 import android.os.Bundle;
 
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 
 public class InjectArticleListActivity extends AppCompatActivity {
 
