@@ -6,7 +6,7 @@ import org.kaerdan.presenterretainer.PresenterActivity;
 
 import android.os.Bundle;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 public class RetainPresenterArticleListActivity extends PresenterActivity
         implements RetainPresenterArticleListContract.NavigatorProvider {
